Reverted client_compat.go into client.go and renamed dialer variable
